feat(pick): make key query process respect the 'iminsert' option
The 'getcharstr' returns a character from the input as is, and it is not
affected by the 'iminsert' option.

It is possible, however, to construct a lookup table for ':lmap'
mappings and apply it to the returned character on the fly.

This successfully emulates the 'iminsert' option in the 'mini.pick'
input field.

However, the 'iminsert' toggle has to be defined in 'mini.pick'
configuration as a keybinding, as the standard '<C-^>' has no effect by
default.

The lookup table is created only once and then cached. In my
non-scientific testing, thie feature has no impact on input latency.

The feature has no effect if the 'keymap' option is not set.

I understand that 'iminsert' is a relatively niche feature to cover, and
it was never mentioned in issues or discussions. While you can always
change the keyboard layout, and achieve the desired effect, it feels
unnatural to me. I write quite a lot in the 'iminsert' mode, and find it
easier to invoke than the keyboard layout change.

I thought to raise the issue beforehand, but the implementation is quite
trivial, so I decided to raise the PR outright.

I intentionally did not update the documentation or anything. If you
want to move forward with the feature, I am ready and willing to follow
your lead and bring the change up to standard.